Create Domain layer
