+++ src/SubtitleFileCleanerGUI.Domain/Attributes/SinglePathAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace SubtitleFileCleanerGUI.Attributes
+namespace SubtitleFileCleanerGUI.Domain.Attributes
 {
     // Stores metadata about file path (single means one attribute per field)
     [AttributeUsage(AttributeTargets.Field, AllowMultiple = false)]
