+++ src/SubtitleFileCleanerGUI.Domain/Attributes/SubtitleCleanerAttribute.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using SubtitleBytesClearFormatting.Cleaners;
 
-namespace SubtitleFileCleanerGUI.Attributes
+namespace SubtitleFileCleanerGUI.Domain.Attributes
 {
     // Stores metadata about the subtitle cleaner type
     [AttributeUsage(AttributeTargets.Field, AllowMultiple = false)]
