+++ src/SubtitleFileCleanerGUI.Domain/SubtitleFileCleanerGUI.Domain.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="CommunityToolkit.Mvvm" Version="7.1.2" />
+    <PackageReference Include="SubtitleBytesClearFormatting" Version="5.0.0" />
+  </ItemGroup>
+
+</Project>
