+++ src/SubtitleFileCleanerGUI.UI/Service/ModelCreation/ISubtitleStatusFileCreator.cs
@@ -1,4 +1,4 @@
-﻿using SubtitleFileCleanerGUI.Model;
+﻿using SubtitleFileCleanerGUI.Domain.Model;
 
 namespace SubtitleFileCleanerGUI.Service.ModelCreation
 {
