+++ src/SubtitleFileCleanerGUI.UI/Service/ModelCreation/StatusInfoWatcher.cs
@@ -1,7 +1,8 @@
 ﻿using System.ComponentModel;
 using System.Linq;
-using SubtitleFileCleanerGUI.Attributes;
-using SubtitleFileCleanerGUI.Model;
+using SubtitleFileCleanerGUI.Domain.Attributes;
+using SubtitleFileCleanerGUI.Domain.Enums;
+using SubtitleFileCleanerGUI.Domain.Model;
 using SubtitleFileCleanerGUI.Service.Utility;
 
 namespace SubtitleFileCleanerGUI.Service.ModelCreation
