+++ src/SubtitleFileCleanerGUI.UI/Service/SubtitleConversion/AutoCleanerDefiner.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Linq;
-using SubtitleFileCleanerGUI.Attributes;
-using SubtitleFileCleanerGUI.Model;
+using SubtitleFileCleanerGUI.Domain.Attributes;
+using SubtitleFileCleanerGUI.Domain.Enums;
 using SubtitleFileCleanerGUI.Service.Utility;
 
 namespace SubtitleFileCleanerGUI.Service.SubtitleConversion
