+++ src/SubtitleFileCleanerGUI.UI/Service/SubtitleConversion/IAutoCleanerDefiner.cs
@@ -1,4 +1,4 @@
-﻿using SubtitleFileCleanerGUI.Model;
+﻿using SubtitleFileCleanerGUI.Domain.Enums;
 
 namespace SubtitleFileCleanerGUI.Service.SubtitleConversion
 {
