+++ src/SubtitleFileCleanerGUI.UI/Service/SubtitleConversion/SubtitleCleanerCreator.cs
@@ -1,7 +1,7 @@
 ﻿using System.Linq;
 using SubtitleBytesClearFormatting.Cleaners;
-using SubtitleFileCleanerGUI.Attributes;
-using SubtitleFileCleanerGUI.Model;
+using SubtitleFileCleanerGUI.Domain.Attributes;
+using SubtitleFileCleanerGUI.Domain.Enums;
 using SubtitleFileCleanerGUI.Service.Utility;
 
 namespace SubtitleFileCleanerGUI.Service.SubtitleConversion
