+++ src/SubtitleFileCleanerGUI.UI/ViewModel/SettingsVM.cs
@@ -3,7 +3,8 @@ using System.Windows;
 using System.Windows.Input;
 using CommunityToolkit.Mvvm.ComponentModel;
 using Microsoft.Extensions.Logging;
-using SubtitleFileCleanerGUI.Model;
+using SubtitleFileCleanerGUI.Domain.Enums;
+using SubtitleFileCleanerGUI.Domain.Model;
 using SubtitleFileCleanerGUI.Service.Input;
 using SubtitleFileCleanerGUI.Service.Settings;
 using SubtitleFileCleanerGUI.Service.Utility;
